backend/requirements.txt
acres==0.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.17.2
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asyncpg==0.31.0
attrs==25.4.0
bcrypt==5.0.0
billiard==4.2.4
black==25.12.0
celery==5.6.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
ci-info==0.3.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
configobj==5.0.9
configparser==7.2.0
cryptography==46.0.3
Deprecated==1.3.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
etelemetry==0.3.1
fastapi==0.128.0
filelock==3.20.3
fitz==0.0.1.dev2
frozenlist==1.8.0
gevent==25.9.1
greenlet==3.3.0
h11==0.16.0
httplib2==0.31.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jwt==1.4.0
kombu==5.6.2
limits==5.6.0
looseversion==1.3.0
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.0
mypy_extensions==1.1.0
networkx==3.6.1
nibabel==5.3.3
nipype==1.10.0
numpy==2.4.1
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pathlib==1.0.1
pathspec==0.12.1
pgvector==0.4.2
pillow==12.1.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
propcache==0.4.1
prov==2.1.1
puremagic==1.30
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydot==4.0.1
pyparsing==3.3.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.21
pytokens==0.3.0
pytz==2025.2
pyxnat==1.6.4
rdflib==7.5.0
redis==6.4.0
requests==2.32.5
rsa==4.9.1
scipy==1.17.0
simplejson==3.20.2
six==1.17.0
slowapi==0.1.9
sqladmin==0.22.0
SQLAlchemy==2.0.45
starlette==0.50.0
tabulate==0.9.0
traits==7.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
unicorn==2.1.4
urllib3==2.6.3
uvicorn==0.40.0
vine==5.1.0
wcwidth==0.2.14
websocket==0.2.1
websockets==16.0
wrapt==2.0.1
WTForms==3.1.2
yarl==1.22.0
zope.event==6.1
zope.interface==8.2
